Remove language enum
