+++ src/main/java/stream/flarebot/flarebot/objects/ReportStatus.java
@@ -1,8 +1,5 @@
 package stream.flarebot.flarebot.objects;
 
-import stream.flarebot.flarebot.FlareBot;
-import stream.flarebot.flarebot.Language;
-
 public enum ReportStatus {
 
     OPEN("Open"),
